+++ es-core/src/main/kotlin/io/provenance/eventstream/net/OkHttpAdapter.kt
@@ -1,8 +1,6 @@
 package io.provenance.eventstream.net
 
 import com.tinder.scarlet.websocket.okhttp.newWebSocketFactory
-import io.provenance.eventstream.WsAdapter
-import io.provenance.eventstream.stream.clients.BlockFetcher
 import io.provenance.eventstream.stream.clients.TendermintBlockFetcher
 import io.provenance.eventstream.stream.clients.TendermintServiceOpenApiClient
 import mu.KotlinLogging
@@ -14,7 +12,7 @@ import kotlin.time.Duration.Companion.seconds
 import kotlin.time.ExperimentalTime
 
 /**
- * Create a default okHttpClient to use for the event stream.
+ * Create a default [OkHttpClient] to use within the event stream.
  */
 @OptIn(ExperimentalTime::class)
 fun defaultOkHttpClient(pingInterval: Duration = 10.seconds, readInterval: Duration = 60.seconds) =
@@ -23,23 +21,26 @@ fun defaultOkHttpClient(pingInterval: Duration = 10.seconds, readInterval: Durat
         .readTimeout(readInterval.inWholeMilliseconds, TimeUnit.MILLISECONDS)
         .build()
 
+/**
+ * Create the [OkHttpClient] flavor of the required [NetAdapter] fields.
+ *
+ * @param hosh The node host address to connect to.
+ * @param okHttpClient The [OkHttpClient] instance to use for http calls.
+ * @param tls Are the connections running over tls?
+ * @return The [NetAdapter] instance.
+ */
 fun okHttpNetAdapter(host: String, okHttpClient: OkHttpClient = defaultOkHttpClient(), tls: Boolean = false): NetAdapter {
     fun scheme(pre: String) = if (tls) "${pre}s" else pre
 
-    return object : NetAdapter {
-        private val log = KotlinLogging.logger {}
+    val log = KotlinLogging.logger {}
+    val wsUri = URI.create("${scheme("ws")}://$host")
+    val rpcUri = URI.create("${scheme("http")}://$host")
 
-        val wsUri = URI.create("${scheme("ws")}://$host")
-        val rpcUri = URI.create("${scheme("http")}://$host")
+    log.info { "initializing ws endpoint:$wsUri" }
+    log.info { "initializing rpc endpoint:$rpcUri" }
 
-        init {
-            log.info { "initializing ws endpoint:$wsUri" }
-            log.info { "initializing rpc endpoint:$rpcUri" }
-        }
-
-        override val wsAdapter: WsAdapter =
-            okHttpClient.newWebSocketFactory("${wsUri.scheme}://${wsUri.host}:${wsUri.port}/websocket")::create
-        override val rpcAdapter: BlockFetcher =
-            TendermintBlockFetcher(TendermintServiceOpenApiClient(rpcUri.toASCIIString()))
-    }
+    return netAdapter(
+        okHttpClient.newWebSocketFactory("${wsUri.scheme}://${wsUri.host}:${wsUri.port}/websocket")::create,
+        TendermintBlockFetcher(TendermintServiceOpenApiClient(rpcUri.toASCIIString()))
+    )
 }
